refactor: ActionsSettings.tsxコンポーネントの分割とコード整理
- ActionsSettings.tsx (550行) → 5つのコンポーネントに分割
  - ActionsHeader: ナビゲーションヘッダー
  - ActionItem: 個別アクション表示コンポーネント
  - ActionForm: アクション編集・作成フォーム
  - ActionsList: ドラッグ&ドロップ対応リスト
  - ResetConfirmDialog: 確認ダイアログ

- 共通ユーティリティを追加
  - contentTypes.ts: コンテンツタイプ定義の統一化
  - 重複していたアイコンマッピングを既存utilsに統一

- コード品質向上
  - 550行 → 136行 (メインファイル75%削減)
  - TypeScript型安全性の強化
  - 単一責任原則の適用
