+++ src/components/actions/ActionsHeader.tsx
@@ -1,0 +1,23 @@
+import { Button } from "@/components/ui/button";
+import { ArrowLeft } from "lucide-react";
+import { useNavigate } from "react-router";
+
+export function ActionsHeader() {
+  const navigate = useNavigate();
+
+  return (
+    <div className="border-b">
+      <div className="flex items-center gap-3 p-4">
+        <Button
+          variant="ghost"
+          size="icon"
+          onClick={() => navigate('/')}
+          className="h-8 w-8"
+        >
+          <ArrowLeft className="h-4 w-4" />
+        </Button>
+        <h1 className="text-lg font-semibold">アクション設定</h1>
+      </div>
+    </div>
+  );
+}
